# justdoit-backend/src/main/java/com/ubt/todo/request/ResetPasswordRequest.java
package com.ubt.todo.request;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class ResetPasswordRequest {

    private String email;
    private String password;
    private String confirmPassword;
}