# justdoit-backend/src/main/java/com/ubt/todo/request/UserAuthenticationRequest.java
package com.ubt.todo.request;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import javax.validation.constraints.NotNull;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class UserAuthenticationRequest {

    @NotNull
    private String username;
    @NotNull
    private String password;
}