# justdoit-frontend/src/app/services/todo-api/swagger.json
{"swagger":"2.0","info":{"description":"TODO API reference for developers","title":"TODO API"},"host":"192.168.1.51:8080","basePath":"/","tags":[{"name":"authentication-controller","description":"Authentication Controller"},{"name":"test-controller","description":"Test Controller"},{"name":"todo-controller","description":"Todo Controller"},{"name":"user-controller","description":"User Controller"},{"name":"user-relationship-controller","description":"User Relationship Controller"}],"paths":{"/api/relationship/all":{"get":{"tags":["user-relationship-controller"],"summary":"getAllPeople","operationId":"getAllPeopleUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserRelationshipTransport"}}}}}},"/api/relationship/friends":{"get":{"tags":["user-relationship-controller"],"summary":"getAllFriends","operationId":"getAllFriendsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserRelationshipTransport"}}}}}},"/api/relationship/pending":{"get":{"tags":["user-relationship-controller"],"summary":"geAllPendingRequests","operationId":"geAllPendingRequestsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserRelationshipTransport"}}}}}},"/api/relationship/requested":{"get":{"tags":["user-relationship-controller"],"summary":"getAllRequestedRequests","operationId":"getAllRequestedRequestsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserRelationshipTransport"}}}}}},"/api/relationship/{id}":{"get":{"tags":["user-relationship-controller"],"summary":"get","operationId":"getUsingGET_1","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserRelationshipTransport"}}}},"post":{"tags":["user-relationship-controller"],"summary":"create","operationId":"createUsingPOST_1","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserRelationshipTransport"}}}},"delete":{"tags":["user-relationship-controller"],"summary":"delete","operationId":"deleteUsingDELETE_1","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"204":{"description":"No Content"}}}},"/api/relationship/{id}/accept":{"put":{"tags":["user-relationship-controller"],"summary":"accept","operationId":"acceptUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserRelationshipTransport"}}}}},"/api/relationship/{id}/cancel":{"delete":{"tags":["user-relationship-controller"],"summary":"cancel","operationId":"cancelUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"204":{"description":"No Content"}}}},"/api/relationship/{id}/reject":{"put":{"tags":["user-relationship-controller"],"summary":"reject","operationId":"rejectUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserRelationshipTransport"}}}}},"/api/todos":{"post":{"tags":["todo-controller"],"summary":"create","operationId":"createUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"todoTransport","description":"todoTransport","required":true,"schema":{"$ref":"#/definitions/TodoTransport"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TodoTransport"}}}}},"/api/todos/assigned":{"get":{"tags":["todo-controller"],"summary":"getAssignedTodos","operationId":"getAssignedTodosUsingGET","produces":["*/*"],"parameters":[{"name":"oldestFirst","in":"query","description":"oldestFirst","required":false,"type":"boolean","default":true}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/TodoTransport"}}}}}},"/api/todos/created":{"get":{"tags":["todo-controller"],"summary":"getCreatedTodos","operationId":"getCreatedTodosUsingGET","produces":["*/*"],"parameters":[{"name":"oldestFirst","in":"query","description":"oldestFirst","required":false,"type":"boolean","default":true}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/TodoTransport"}}}}}},"/api/todos/{todoId}":{"get":{"tags":["todo-controller"],"summary":"get","operationId":"getUsingGET","produces":["*/*"],"parameters":[{"name":"todoId","in":"path","description":"todoId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TodoTransport"}}}},"put":{"tags":["todo-controller"],"summary":"update","operationId":"updateUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"todoId","in":"path","description":"todoId","required":true,"type":"integer","format":"int64"},{"in":"body","name":"todoTransport","description":"todoTransport","required":true,"schema":{"$ref":"#/definitions/TodoTransport"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TodoTransport"}}}},"delete":{"tags":["todo-controller"],"summary":"delete","operationId":"deleteUsingDELETE","produces":["*/*"],"parameters":[{"name":"todoId","in":"path","description":"todoId","required":true,"type":"integer","format":"int64"}],"responses":{"204":{"description":"No Content"}}}},"/api/todos/{todoId}/finished":{"put":{"tags":["todo-controller"],"summary":"markFinished","operationId":"markFinishedUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"todoId","in":"path","description":"todoId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TodoTransport"}}}}},"/api/todos/{todoId}/not-finished":{"put":{"tags":["todo-controller"],"summary":"markNotFinished","operationId":"markNotFinishedUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"todoId","in":"path","description":"todoId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/TodoTransport"}}}}},"/auth/login":{"post":{"tags":["authentication-controller"],"summary":"login","operationId":"loginUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"userAuthenticationRequest","description":"userAuthenticationRequest","required":true,"schema":{"$ref":"#/definitions/UserAuthenticationRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/AuthenticationSuccess"}}}}},"/auth/refresh-token/{token}":{"get":{"tags":["authentication-controller"],"summary":"refreshToken","operationId":"refreshTokenUsingGET","produces":["*/*"],"parameters":[{"name":"token","in":"path","description":"token","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/RefreshTokenSuccess"}}}}},"/test":{"get":{"tags":["test-controller"],"summary":"validateResetToken","operationId":"validateResetTokenUsingGET","produces":["*/*"],"responses":{"204":{"description":"No Content"}}}},"/users/forgot-password":{"post":{"tags":["user-controller"],"summary":"forgotPassword","operationId":"forgotPasswordUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"forgotPasswordRequest","description":"forgotPasswordRequest","required":true,"schema":{"$ref":"#/definitions/ForgotPasswordRequest"}}],"responses":{"204":{"description":"No Content"}}}},"/users/reset-password/{token}":{"get":{"tags":["user-controller"],"summary":"validateResetToken","operationId":"validateResetTokenUsingGET_1","produces":["*/*"],"parameters":[{"name":"token","in":"path","description":"token","required":true,"type":"string"}],"responses":{"204":{"description":"No Content"}}},"post":{"tags":["user-controller"],"summary":"resetUserPassword","operationId":"resetUserPasswordUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"resetPasswordRequest","description":"resetPasswordRequest","required":true,"schema":{"$ref":"#/definitions/ResetPasswordRequest"}},{"name":"token","in":"path","description":"token","required":true,"type":"string"}],"responses":{"204":{"description":"No Content"}}}},"/users/sign-up":{"post":{"tags":["user-controller"],"summary":"registerAccount","operationId":"registerAccountUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"registrationRequest","description":"registrationRequest","required":true,"schema":{"$ref":"#/definitions/RegistrationRequest"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserTransport"}}}}},"/users/user-confirmation/{token}":{"get":{"tags":["user-controller"],"summary":"userConfirmation","operationId":"userConfirmationUsingGET","produces":["*/*"],"parameters":[{"name":"token","in":"path","description":"token","required":true,"type":"string"}],"responses":{"204":{"description":"No Content"}}}}},"securityDefinitions":{"Bearer %token":{"type":"apiKey","name":"Authorization","in":"header"}},"definitions":{"AuthenticationSuccess":{"type":"object","properties":{"authToken":{"type":"string"},"email":{"type":"string"},"reason":{"type":"string","enum":["AUTHENTICATED","REFRESHED_TOKEN"]},"refreshToken":{"type":"string"},"success":{"type":"boolean"},"userId":{"type":"integer","format":"int64"},"username":{"type":"string"}},"title":"AuthenticationSuccess"},"ForgotPasswordRequest":{"type":"object","required":["email"],"properties":{"email":{"type":"string"}},"title":"ForgotPasswordRequest"},"RefreshTokenSuccess":{"type":"object","properties":{"authToken":{"type":"string"},"email":{"type":"string"},"reason":{"type":"string","enum":["AUTHENTICATED","REFRESHED_TOKEN"]},"refreshToken":{"type":"string"},"success":{"type":"boolean"}},"title":"RefreshTokenSuccess"},"RegistrationRequest":{"type":"object","required":["confirmPassword","email","password","username"],"properties":{"confirmPassword":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"username":{"type":"string"}},"title":"RegistrationRequest"},"ResetPasswordRequest":{"type":"object","properties":{"confirmPassword":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"title":"ResetPasswordRequest"},"TodoTransport":{"type":"object","required":["assignee","description","dueDate","title"],"properties":{"assignee":{"$ref":"#/definitions/UserTransport"},"createdAt":{"type":"string","format":"date-time"},"createdBy":{"$ref":"#/definitions/UserTransport"},"description":{"type":"string"},"dueDate":{"type":"string","format":"date-time"},"finishedAt":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"status":{"type":"string","enum":["CREATED","EXPIRED","FINISHED"]},"title":{"type":"string"},"updatedAt":{"type":"string","format":"date-time"}},"title":"TodoTransport"},"UserAuthenticationRequest":{"type":"object","required":["password","username"],"properties":{"password":{"type":"string"},"username":{"type":"string"}},"title":"UserAuthenticationRequest"},"UserRelationshipTransport":{"type":"object","properties":{"relatedUser":{"$ref":"#/definitions/UserTransport"},"requestedAt":{"type":"string","format":"date-time"},"userRelationshipType":{"type":"string","enum":["BLOCKED","FRIENDS","NONE","PENDING"]}},"title":"UserRelationshipTransport"},"UserTransport":{"type":"object","properties":{"email":{"type":"string"},"id":{"type":"integer","format":"int64"},"username":{"type":"string"}},"title":"UserTransport"}}}